% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{color}
\alias{color}
\alias{colors}
\title{Change font color for multiple Rmd output types}
\usage{
color(..., color)

colors()
}
\arguments{
\item{...}{Any number of strings (or objects coercible to strings via
\code{\link[=paste0]{paste0()}}), which will be pasted together via \code{\link[=paste0]{paste0()}} with font color
specified in \code{color}.}

\item{color}{Color to apply to font.}
}
\value{
A formatted string, appropriate for the current output type.
}
\description{
Apply specified font color to specified text, formatted appropriately for
different Rmd outputs. RGB colors of the UIUC brand are supplied via
\code{\link[=colors]{colors()}}.
}
\examples{
color("hello world", color = latexWordHelpers::colors()$IlliniOrange)
}
