% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{itemize}
\alias{itemize}
\alias{item}
\alias{enumerate}
\title{Make lists for multiple Rmd outputs}
\usage{
itemize(x = 0)

item(..., y = NULL, l = 1)

enumerate(..., start = 1)
}
\arguments{
\item{x}{Set to \code{0} (the default) to begin an unordered list, and to \code{1} to
end an unordered list via \code{\link[=itemize]{itemize()}}.}

\item{...}{Any number of strings or objects coercible to strings via
\code{\link[=paste]{paste()}}. For \code{\link[=itemize]{itemize()}}, these inputs will be combined
into a single item in the output with an empty string as the separator. For
\code{\link[=enumerate]{enumerate()}}, each argument to \code{...} becomes one item in
the list.}

\item{y}{Options to pass to item in latex output (i.e. \item[y]{...}).}

\item{l}{Depth of item in (nested) lists. Defaults to 1. Only necessary if
there is a non-latex output, as item depth is determined via the number of
tab separators preceding the item.}

\item{start}{Number at which to start the ordered list.}
}
\value{
A formatted string appropriate for the current output type.
}
\description{
Compact functions to make lists in multiple Rmd output formats.
}
\examples{
# Unordered list, with one item's symbol changed to a diamond
itemize()
item("Hello world")
item("Hello world", "2", y = "\\\\$\\\\diamond\\\\$")
itemize(1)

# Unordered, nested lists
itemize() # start level 1 list
item("Hello world")
itemize() # start level 2 list
item("Hello world", "2", y = "\\\\$\\\\diamond\\\\$", l = 2)
itemize(1) # end level 2 list
itemize(1) # end level 1 list

# Ordered list
enumerate("One", "Two", "Three")

# Ordered list starting at 2
enumerate("One", "Two", "Three", start = 2)
}
