% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltx.R
\name{ltx}
\alias{ltx}
\title{Send text to latex output}
\usage{
ltx(...)
}
\arguments{
\item{...}{Any number of strings or objects coercible to strings.}
}
\value{
If the current output is latex and arguments are passed to \code{...},
returns a string. If the current output is not latex and arguments are
passed to \code{...}, returns an empty string. If the current output is latex
and arguments are not passed to \code{...}, returns \code{TRUE}. If the current
output is latex and arguments are not passed to \code{...}, returns \code{FALSE}.
}
\description{
Pass any number of strings (or objects coercible to strings via \code{\link[=paste]{paste()}}),
and they will be printed to latex output only. Note that arguments will be
collapsed with an empty string as the separator.
}
\details{
If no arguments are passed to the function, it becomes a logical test
equivalent to \code{\link[knitr:output_type]{knitr::is_latex_output()}}.
}
\examples{
\dontrun{
  cat(ltx("hello", " world"))
  if (ltx()) cat("\\\\textbf{hello world}")
}

}
\seealso{
\code{\link[latexWordHelpers:word]{latexWordHelpers::word()}}
}
