\name{rmdHelpers-package}
\alias{rmdHelpers-package}
\alias{rmdHelpers}
\docType{package}
\title{
\packageTitle{rmdHelpers}
}
\description{
\packageDescription{rmdHelpers}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rmdHelpers}
\packageIndices{rmdHelpers}
Basic functions that I use in multiple Rmd documents
}
\author{
\packageAuthor{rmdHelpers}

Maintainer: \packageMaintainer{rmdHelpers}
}

\keyword{ package }
