% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{br}
\alias{br}
\alias{newpage}
\title{Insert space characters with one function for multiple Rmd output types}
\usage{
br(x = 1)

newpage()
}
\arguments{
\item{x}{If output is latex, this argument specifies which type of line break
to use: \code{0} for \verb{\\newline} and \code{1} for \verb{\\linebreak} (the default). If
output is not latex this argument is ignored.}
}
\value{
A formatted string, appropriate for current output type
}
\description{
Add linebreaks or new pages to multiple Rmd output types with compact
functions.
}
\examples{
cat(br(1))
cat(br(0))
cat(newpage())
}
