% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{word}
\alias{word}
\title{Send text to Word output}
\usage{
word(...)
}
\arguments{
\item{...}{Any number of strings or objects coercible to strings.}
}
\value{
If the current output is Word and arguments are passed to \code{...},
returns a string. If the current output is not Word and arguments are
passed to \code{...}, returns an empty string. If the current output is Word
and arguments are not passed to \code{...}, returns \code{TRUE}. If the current
output is Word and arguments are not passed to \code{...}, returns \code{FALSE}.
}
\description{
Pass any number of strings (or objects coercible to strings via \code{\link[=paste]{paste()}}),
and they will be printed to Word output only. Note that arguments will be
collapsed with an empty string as the separator.
}
\details{
If no arguments are passed to the function, it becomes a logical test
evaluating if the output is \emph{not} latex.
}
\examples{
\dontrun{
  cat(word("hello", " ", "world"))
  if (word()) cat("**hello world**")
}

}
\seealso{
\code{\link[latexWordHelpers:ltx]{latexWordHelpers::ltx()}}
}
