% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontface.R
\name{bf}
\alias{bf}
\alias{it}
\alias{bi}
\alias{sb}
\title{Easy fontface changes in multiple Rmd outputs}
\usage{
bf(x)

it(x)

bi(x)

sb(x)
}
\arguments{
\item{x}{A string or an object coercible to a string}
}
\value{
A formatted string appropriate for the current output
}
\description{
Shorthand functions to add a fontface when there are multiple Rmd output
types. For bold face: \code{\link[=bf]{bf()}}, for italic:
\code{\link[=it]{it()}}, for bold-italic \code{\link[=bi]{bi()}}, for
semibold \code{\link[=sb]{sb()}}.
}
\examples{
bf("hello world")
it("hello world")
bi("hello world")
sb("hello world")
}
